feat(minor-multisig-prover): remove external gateway encoding v1
- remove deprecated types
- remove abi encoding
- remove bsc encoding, will add the new bsc encoding when the new external Sui Gateway is ready
