+++ contracts/multisig-prover/src/types.rs
@@ -1,40 +1,10 @@
-use std::fmt::Display;
-
 use axelar_wasm_std::{Participant, Snapshot};
 use cosmwasm_schema::cw_serde;
-use cosmwasm_std::{from_binary, HexBinary, StdResult, Uint256};
+use cosmwasm_std::{from_binary, HexBinary, StdResult};
 use cw_storage_plus::{Key, KeyDeserialize, PrimaryKey};
-use multisig::{
-    key::{PublicKey, Signature},
-    worker_set::WorkerSet,
-};
+use multisig::{key::PublicKey, worker_set::WorkerSet};
 use router_api::CrossChainId;
 use sha3::{Digest, Keccak256};
-
-use crate::encoding::{Data, Encoder};
-
-#[cw_serde]
-pub enum CommandType {
-    ApproveContractCall,
-    TransferOperatorship,
-}
-
-impl Display for CommandType {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            CommandType::ApproveContractCall => write!(f, "approveContractCall"),
-            CommandType::TransferOperatorship => write!(f, "transferOperatorship"),
-        }
-    }
-}
-
-#[deprecated(since = "0.2.0", note = "Please use payload module instead")]
-#[cw_serde]
-pub struct Command {
-    pub id: HexBinary,
-    pub ty: CommandType,
-    pub params: HexBinary,
-}
 
 #[cw_serde]
 pub struct BatchId(HexBinary);
@@ -82,37 +52,6 @@ impl BatchId {
             message_ids.push(new_worker_set.hash().to_string())
         }
         Keccak256::digest(message_ids.join(",")).as_slice().into()
-    }
-}
-
-#[deprecated(since = "0.2.0", note = "Please use payload module instead")]
-#[cw_serde]
-pub struct CommandBatch {
-    pub id: BatchId,
-    pub message_ids: Vec<CrossChainId>,
-    pub data: Data,
-    pub encoder: Encoder,
-}
-
-#[deprecated(
-    since = "0.2.0",
-    note = "Please use multisig::msg::SignerWithSig instead"
-)]
-#[cw_serde]
-#[derive(Ord, PartialOrd, Eq)]
-pub struct Operator {
-    pub address: HexBinary,
-    pub weight: Uint256,
-    pub signature: Option<Signature>,
-}
-
-impl Operator {
-    pub fn with_signature(self, sig: Signature) -> Operator {
-        Operator {
-            address: self.address,
-            weight: self.weight,
-            signature: Some(sig),
-        }
     }
 }
